Make amends to DomPatch objects, and how they're added to the history stack.
- Make DomPatch objects have two stringPos values; `backStart` and
`forwardStart`, recognizing that depending on if we're using the patch
list for undoing or redoing, string positions of nodes in the active
document HTML string will be different.

- Create two maps, `stringPosBackwardUpdateMap` and
`stringPosForwardUpdateMap`, for use by HistoryManager in mapping the
stringPos values of DomPatch objects to the appropriate version of the
HTML string.

- Make sure stringPos values in DomPatch objects are index shifted along
with the docNodes entries in the relevant `handle*` functions

- Update the logic of `addToHistoryStack`.

- Add an extensive top-level comment to`undo.ts` which walks through the
logic of using the DomPatches.