Perform more fixes for basic undo/redo functionality
- Fix spurious matches of forwardStart values in `redo()` by creating a
new boolean field `mapForwardStart` on the `DomPatch` type, used to
determine whether or not the corresponding `forwardStart` value was able
to be properly mapped to the "previous" HTML string (that is the current
HTML string when `redo()` is called) or not.

- Fix the handling of adding nodes at the top of the document in
`undo()` and `redo()` by handling `*Start` values of `-1`.

- Fix the `prevStrPosForwardMap` passed into `mapForwardStartVals` being
incomplete when adding a new history item (due to early returns when
constructing the maps).

- Fix `handleNodeRemove()` sometimes finding the wrong previous node,
and simplify the logic (to search directly the previous node, rather
than sort nodes by how "directly previous" they are).

- Comment out integration of old undo/redo logic for the CSS editor.

- Improve logging for the history manager.

- Add some musings for the history manager in `Guidance Document
Editor.md`.