Make large changes to core undo/redo, fixing issues
- Change to using an XML parser for undo/redo, in order to avoid
corrections to structure present when using a HTML parser.

- Change from encoding HTML strings using named entity references to
only using numerical references, and change some aspects of
`getElementString` in docsyncing to support an XML parser.

- Fix ephemeralNodeMap entries not being added for ephemeral nodes which
are implicitly added and/or removed, by adding an extra field `subPos`
to `DocPatchDom`, filled in in docsyncing with offsets of the HTML
string used for adding nodes, for the starts and ends of all said nodes.
This is then utilised by a new method `handleEphemeralAddOrRemovePatch`
on HistoryManager.

- Fix nodes which were added in through undo/redo having adjacent child
text nodes be interpreted as a single text node (not expected by patch
lists), by adding an extra field `expectedLen` to `DocPatchDom`, filled
in by docsyncing and utilised by the undo/redo methods to split text
nodes at runtime when it detects this.

- Fix ephemeral nodes which were only added in in the exact previous
HTML string being spuriously matched, by keeping track of an extra list
of numbers `prevNodeAddLocs` on HistoryManager.

- Fix the handling of forwardStarts on the newly added patches (by
realising it's only possible for remove-type patches to have
forwardStarts referring to ephemeral nodes).

- Fix the `ephemeralNodeForwardMap` not being properly populated, by
moving the logic out of `handleForwardMappedBackStarts`.

- Fix the logic of how the `ephemeralNodBackMap` was utilised (by
realising that each entry will point to a real node straight away- no
chaining backStarts into the map required).

- Fix errors while undoing/redoing for unresolved stringPos values using
posNodes when dealing with patches for ephemeral nodes, by placing the
checks for skipping the patch before attempting to find an entry in
posNodes.

- Improve comments in `history.ts`.