Improve relability of undo/redo, and document loading
- Change primary document loading to use an XML parser (avoiding
potential structure corrections from HTML parsers).

- Add error catching to the `undo` and `redo` methods in `App.svelte`,
to reset the HTML string of the document upon failure midway through
undoing/redoing (i.e. making undo/redo operations atomic).

- Fix entries for added child nodes in `tempPosNodes` not being properly
mapped to a version of the HTML string (the base version when undoing,
and the latest version when redoing), by creating a new field
`subPosMapped` on `DocPatchDom`, that describes string positions that
are kept mapped to the base or latest HTML string, and having
`HistoryManager.addNode` use `subPosMapped` instead of
`reconstructHTMLString()` (and also handle combined text nodes).

- Fix `handleEphemeralAddOrRemovePatch` to now also modify the passed in
`subOffsets` (formerly subPos)― and `subPosMapped`― to remove references
to child ephemeral nodes.

- Fix patches referring to nodes added and then removed in successive
sets of patches that are merged in ("quick" ephemeral nodes) not having
their `backStart` set to -2.

- Fix domPatches in `handleAttributeChange` in docsyncing not having
stringPos values index shifted when they referred to nodes that were
children of the modified element.

- Fix empty HTML strings being treated as missing in `docsyncing`.

- Remove other logic for handling combined text nodes, as they are now
dealt with immediately in `HistoryManager.addNode`.

- Factor out the `parseHTMLFragment` function to `helper.ts`, and
improve reliability by having a fallback to a HTML parser on XML parse
errors.

- Factor out `he.encode` and `he.decode` into a single "source of truth"
for encoding and decoding HTML text, into `encodeHTMLText` and
`decodeHTMLText` in `helper.ts`.

- Make adding and removing nodes suggest new history items.

- Change the handling of newly added nodes according to 2) of the
MutationObserver in `App.svelte`, to not add NBSPs to empty text nodes
(i.e. now it's only to empty elements).

- Improve documentation by adding a brief glossary of terms in the
beginning of `history.ts`, and a docstring to
`HistoryManager.handleEphemeralAddOrRemovePatch`.