Add pruning to local save entries & Add max history stack size
- Send any queued patches to the server on page leave using the
"visibilitychange" event and `keepalive: true` on the fetch call.

- Implement a maximum history stack size (set to 100 for now).

- Add exported function `pruneLocalSave` in `store.svelte.ts` which
utilises the "lastUsed" field to prune entries after not being used for
a certain amount of time (currently 7 days), for now only calling this
function whenever the local save of a document is updated.

- Fix `generateCompContentStr` inserting "data-component" attributes in
the wrong location in the HTML string for void elements.

- Fix `stringPosNodeMap` not being generated for the initial document
state (after page load), in `docsaving.svelte.ts`.

- Fix condition in `HistoryManager.undo()` allowing undo when at the
base state.